Edits are now in the middleware flow
