+++ lib/telegram2discord/middlewares.js
@@ -9,6 +9,8 @@ const Bridge = require("../bridgestuff/Bridge");
 const From = require("./From");
 const mime = require("mime/lite");
 const request = require("request");
+const handleEntities = require("./handleEntities");
+const Discord = require("discord.js");
 
 /***********
  * Helpers *
@@ -41,6 +43,56 @@ function createTextObjFromMessage(message) {
 		// Default to undefined
 		[R.T, R.always({ raw: "", entities: [] })]
 	])(message);
+}
+
+/**
+ * Makes the reply text to show on Discord
+ *
+ * @param {Object} replyTo	The replyTo object from the tediCross context
+ * @param {Integer} replyLength	How many characters to take from the original
+ * @param {Integer} maxReplyLines	How many lines to cut the reply text after
+ *
+ * @returns {String}	The reply text to display
+ */
+function makeReplyText(replyTo, replyLength, maxReplyLines) {
+	// Make the reply string
+	return R.compose(
+		// Add ellipsis if the text was cut
+		R.ifElse(
+			R.compose(
+				R.equals(R.length(replyTo.text.raw)),
+				R.length
+			),
+			R.identity,
+			R.concat(R.__, "…")
+		),
+		// Take only a number of lines
+		R.join("\n"),
+		R.slice(0, maxReplyLines),
+		R.split("\n"),
+		// Take only a portion of the text
+		R.slice(0, replyLength),
+	)(replyTo.text.raw);
+}
+
+/**
+ * Makes a discord mention out of a username
+ *
+ * @param {String} username	The username to make the mention from
+ * @param {Discord.Client} dcBot	The Discord bot to look up the user's ID with
+ * @param {String} channelId	ID of the Discord channel to look up the username in
+ *
+ * @returns {String}	A Discord mention of the user
+ */
+function makeDiscordMention(username, dcBot, channelId) {
+	// Get the name of the Discord user this is a reply to
+	const dcUser = dcBot.channels.get(channelId).members.find(R.propEq("displayName", username));
+
+	return R.ifElse(
+		R.isNil,
+		R.always(username),
+		dcUser => `<@${dcUser.id}>`
+	)(dcUser);
 }
 
 /****************************
@@ -406,18 +458,96 @@ function addFileStream(ctx, next) {
 		.then(R.always(undefined));
 }
 
-/**
- * Adds a boolean telling if the message is an edit 
- *
- * @param {Object} ctx	The context to add the property to
- * @param {Object} ctx.tediCross	The tedicross object on the context
- * @param {Object} ctx.tediCross.message	The message being handled
- * @param {Function} next	Function to pass control to next middleware
- *
- * @returns {undefined}
- */
-function addEditBool(ctx, next) {
-	ctx.tediCross.isEdit = !R.isNil(ctx.tediCross.message.edit_date);
+function addPreparedObj(ctx, next) {
+	// Shorthand for the tediCross context
+	const tc = ctx.tediCross;
+
+	ctx.tediCross.prepared = R.map(
+		bridge => {
+			// Make the header
+			const header = R.ifElse(
+				R.always(bridge.telegram.sendUsernames),
+				tc => {
+					// Get the name of the sender of this message
+					const senderName = From.makeDisplayName(ctx.TediCross.settings.telegram.useFirstNameInsteadOfUsername, tc.from);
+
+					// Make the default header
+					let header = `**${senderName}**`;
+
+					if (!R.isNil(tc.replyTo)) {
+						// Add reply info to the header
+						const repliedToName = R.ifElse(
+							R.prop("isReplyToTediCross"),
+							R.compose(
+								username => makeDiscordMention(username, ctx.TediCross.dcBot, bridge.discord.channelId),
+								R.prop("dcUsername")
+							),
+							R.compose(
+								R.partial(From.makeDisplayName, [ctx.TediCross.settings.telegram.useFirstNameInsteadOfUsername]),
+								R.prop("originalFrom")
+							)
+						)(tc.replyTo);
+
+
+						header = `**${senderName}** (in reply to **${repliedToName}**`;
+
+						if (ctx.TediCross.settings.discord.displayTelegramReplies === "inline") {
+							// Make the reply text
+							const replyText = makeReplyText(tc.replyTo, ctx.TediCross.settings.discord.replyLength, ctx.TediCross.settings.discord.maxReplyLines);
+
+							// Put the reply text in the header, replacing newlines with spaces
+							header = `${header}: _${R.replace(/\n/g, " ", replyText)}_)`;
+						} else {
+							// Append the closing parenthesis
+							header = `${header})`;
+						}
+					} else if (!R.isNil(tc.forwardFrom)) {
+						// Handle forwards
+						const origSender = From.makeDisplayName(ctx.TediCross.settings.telegram.useFirstNameInsteadOfUsername, tc.forwardFrom);
+						header = `**${origSender}** (forwarded by **${senderName}**)`;
+					}
+					return header;
+				},
+				R.always("")
+			)(tc);
+
+			// Handle embed replies
+			const embed = R.ifElse(
+				tc => !R.isNil(tc.replyTo) && ctx.TediCross.settings.discord.displayTelegramReplies === "embed",
+				tc => {
+					// Make the text
+					const replyText = handleEntities(tc.replyTo.text.raw, tc.replyTo.text.entities, ctx.TediCross.dcBot, bridge);
+
+					return new Discord.RichEmbed({
+						// Discord will not accept embeds with more than 2048 characters
+						description: R.slice(0, 2048, replyText)
+					});
+				},
+				R.always(undefined)
+			)(tc);
+
+			// Handle file
+			const file = R.ifElse(
+				R.compose(
+					R.isNil,
+					R.prop("file")
+				),
+				R.always(undefined),
+				tc => new Discord.Attachment(tc.file.stream, tc.file.name)
+			)(tc);
+
+			// Make the text to send
+			const text = handleEntities(tc.text.raw, tc.text.entities, ctx.TediCross.dcBot, bridge);
+
+			return {
+				bridge,
+				header,
+				embed,
+				file,
+				text
+			};
+		}
+	)(tc.bridges);
 
 	next();
 }
@@ -442,5 +572,5 @@ module.exports = {
 	addTextObj,
 	addFileObj,
 	addFileStream,
-	addEditBool
+	addPreparedObj
 };
