+++ lib/telegram2discord/setup.js
@@ -89,14 +89,10 @@ function setup(logger, tgBot, dcBot, messageMap, bridgeMap, settings) {
 			tgBot.use(middlewares.addTextObj);
 			tgBot.use(middlewares.addFileObj);
 			tgBot.use(middlewares.addFileStream);
-			tgBot.use(middlewares.addEditBool);
-
-// TODO Temporary logging to see what's going on. Remove before release
-			tgBot.use((ctx, next) => {console.log(ctx.tediCross); next();});
+			tgBot.use(middlewares.addPreparedObj);
 
 			// Apply endwares
-			tgBot.on("edited_message", endwares.handleEdits);
-			tgBot.on("edited_channel_post", endwares.handleEdits);
+			tgBot.on(["edited_message", "edited_channel_post"], endwares.handleEdits);
 			tgBot.use(endwares.relayMessage);
 		})
 		// Start getting updates
